requirements-dev.txt
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
Click==8.1.3
cryptography==38.0.4
decorator==5.1.1
future==0.18.2
icinga2apic==0.7.5
idna==3.4
pycparser==2.21
python-telegram-bot==13.15
PyYAML==6.0
requests==2.28.1
simplejson==3.18.0
six==1.16.0
tornado<6.2,>=6.1
urllib3==1.26.13
pytest~=7.2.0
mock==5.0.1
pytest-mock==3.10.0
